fix(security): zeroize long-lived secrets across modules
- handshake: wipe ephemeral X25519 secret, derived tx/rx session keys,
  the local copy of the Ed25519 identity SK, and all intermediate
  key-derivation buffers in destructors and after use
- messaging: add a SecretGuard RAII type so the per-message AEAD key,
  ephemeral DH secret, and ratchet-mixed key are wiped on every exit
  path (classical and KEM encrypt/decrypt), including the error and
  exception returns that previously leaked them
- audit_logger: wipe the Ed25519 signing key on teardown and the
  transient load buffer after copy
- replay_db: wipe MAC and metadata-encryption keys in a new destructor
- also lands two audit stragglers already in the tree and CHANGELOG:
  the DR nonce-shift UB fix (double_ratchet.hpp) and the ReplayDB
  legacy dead-branch removal

+++ src/double_ratchet.hpp
@@ -327,7 +327,7 @@ inline RatchetMessage DoubleRatchet::encrypt_message(const std::vector<uint8_t>&
     // Derive nonce from message count and key
     std::array<uint8_t, 8> counter_bytes{};
     for (int i = 0; i < 8; ++i) {
-        counter_bytes[i] = static_cast<uint8_t>((m.message_count >> (8 * i)) & 0xFF);
+        counter_bytes[i] = static_cast<uint8_t>((static_cast<uint64_t>(m.message_count) >> (8 * i)) & 0xFF);
     }
     crypto_generichash(m.nonce.data(), m.nonce.size(),
         counter_bytes.data(), counter_bytes.size(),
